22-04

+++ 01-basic/Program4.java
@@ -1,10 +1,11 @@
 import java.util.Scanner;
 
+//factorial
 public class Program4 {
-    public static int factorial(int n){
-        int fact = 1;
-        for(int i=1;i<=n;i++){
-            fact = fact*i;
+    public static long factorial(int n) {
+        long fact = 1;
+        for (int i = 1; i <= n; i++) {
+            fact = fact * i;
         }
         return fact;
     }
@@ -14,9 +15,8 @@ public class Program4 {
         Scanner sc = new Scanner(System.in);
         int num = sc.nextInt();
         sc.close();
-        int result = factorial(num);
-        System.out.println("Factorial of "+num+" is:"+result);
+        long result = factorial(num);
+        System.out.println("Factorial of " + num + " is:" + result);
 
-        
     }
 }
